+++ test/index.test.ts
@@ -1,2 +1,3 @@
 import './time.test';
 import './param.test';
+import './number.test';